需求[#21556]:SR-20200914-01-09 视频B接口开发-NVR存储状态获取接口开发
    [描述说明]: 视频B接口开发-NVR存储状态获取接口开发
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/service/TStdRegionService.java
@@ -30,7 +30,12 @@ public class TStdRegionService{
     @Logs(title = "删除", code = "module",content = "根据页面传入的参数删除数据")
     @Transactional(rollbackFor = Exception.class)
     public int deleteByPrimaryId(Long regionId) {
-        return this.tStdRegionDao.deleteByPrimaryId(regionId);
+        List<Long> list = tStdRegionDao.selectDownId(regionId);
+        if(list != null && list.size() > 0){
+            return tStdRegionDao.batchDelete(list);
+        }else {
+            return this.tStdRegionDao.deleteByPrimaryId(regionId);
+        }
     }
 
     @Logs(title = "更新", code = "module",content = "根据页面传入的参数修改数据")
